✨ add setting modal entry point on the "rule" page
