+++ src/components/homepage.tsx
@@ -7,10 +7,12 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Image from "next/image";
 import Link from "next/link";
 import { useEffect, useState } from "react";
-import SettingsModal from "./modals/settingsModal";
 
-export default function HomePage() {
-  const [showSettingsModal, setShowSettingsModal] = useState(false);
+type HomePageProps = {
+  handleSettingsButtonClick: () => void;
+};
+
+export default function HomePage({ handleSettingsButtonClick }: HomePageProps) {
   const [showTips, setShowTips] = useState(false);
 
   useEffect(() => {
@@ -19,14 +21,6 @@ export default function HomePage() {
       localStorage.setItem("tipsShown", "true");
     }
   }, []);
-
-  function handleCloseSettingsModal() {
-    setShowSettingsModal(false);
-  }
-
-  function handleSettingsButtonClick() {
-    setShowSettingsModal(true);
-  }
 
   return (
     <>
@@ -72,9 +66,6 @@ export default function HomePage() {
           Jouer <FontAwesomeIcon icon={faGamepad} className="ml-2 text-3xl" />
         </Link>
       </div>
-      {showSettingsModal && (
-        <SettingsModal onClose={handleCloseSettingsModal} />
-      )}
     </>
   );
 }
