+++ src/utils/constants.ts
@@ -1,6 +1,6 @@
 export const DEFAULT_NUMBER_OF_WORDS_TO_PICK = 40;
 export const DEFAULT_ROUND_DURATION = 45;
-export const DEFAULT_IS_TIME_PENALTY_FEATURE_ENABLED = true;
+export const DEFAULT_IS_TIME_PENALTY_FEATURE_ENABLED = false;
 export const TIME_PENALTY = 5;
 export const GITHUB_REPO_URL = "https://github.com/Spookywy/time2guess";
-export const APP_VERSION = "1.4.1";
+export const APP_VERSION = "1.5.0";
